% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{blast-class}
\alias{blast-class}
\title{The object class returned by \code{\link{blast}}}
\description{
A multi-item list with the following named values:
\itemize{
 \item{$alignments: }{Best scored alignments with the reference database.}
 \item{$cp: }{Copy number table for 16S genes in the reference database.}
 \item{$parameters: }{Parameters used for the BLAST alignment.}
}
}
\seealso{
\code{\link{blast}}
}
