% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast}
\alias{blast}
\title{BLAST sequences against a reference database}
\usage{
blast(
  sequences,
  blast_output = NULL,
  region = NULL,
  ref_db = NULL,
  ref_cp = NULL,
  max_target_seqs = rexmap_option("blast_max_seqs"),
  word_size = rexmap_option("blast_word_size"),
  verbose = rexmap_option("verbose"),
  debug = FALSE,
  show_args = F,
  output_error = F,
  show_alignment = F,
  ncpu = rexmap_option("ncpu"),
  temp_dir = tempdir()
)
}
\arguments{
\item{sequences}{(Required) Either a FASTA file, a character vector of DNA sequences, or
an abundance table (output from \code{\link{sequence_abundance}})}

\item{region}{Hyper-variable region. If this is not NULL, \code{ref_db} is ignored. Possible
values: 'V4', 'V3-V4'.}

\item{ref_db}{Full path to a custom database.}

\item{ref_cp}{Full path to a copy number table for a custom database.}

\item{max_target_seqs}{Maximum number of target sequences to return for each query.
Note that the temporary output file can be large if this is set > 1000.}

\item{word_size}{Word size used for BLAST alignment.}
}
\description{
Aligns each sequence from the FASTA file, a character vector of DNA sequences,
or an abundance table, against a reference database.
Either \code{region} or both \code{ref_db} and \code{ref_cp} arguments must be specificed.
If \code{region}
is specified, then the \code{ref_db} and \code{ref_cp} arguments are ignored and a
reference database is chosen
from the pre-computed set for that hyper-variable region. For available databases
see columns DB and table in \code{rexmap_option('blast_dbs')}.
This function returns a blast-class object (named list; see ?blast-class for info).
}
