% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse sequences that are exact matches, up to shifts and/or length}
\usage{
collapse(
  ab_in,
  verbose = himap_option("verbose"),
  ncpu = himap_option("ncpu"),
  temp_dir = tempdir(),
  ws_scale = 0.8,
  max_target_seqs = 50,
  copy_table = TRUE
)
}
\description{
Like \code{\link{dada2::collapseNoMismatch}} but a lot faster, multithreaded
and scales well to a very large number of unique sequences (100,000 and more).
}
