% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{dada_denoise}
\alias{dada_denoise}
\title{DADA2 denoising}
\usage{
dada_denoise(fastq_trimmed, fastq_untrimmed, pvalue = 1e-04,
  pvalue_adjusted = NULL, save_intermediate = TRUE,
  use_intermediate = FALSE, multithread = himap_option("ncpu"),
  verbose = T, error_estimation_nsamples = 3)
}
\arguments{
\item{pvalue}{P-value threshold before Bonferroni correction for DADA2.}

\item{pvalue_adjusted}{P-value threshold after Bonferroni correction for DADA2.
If left as NULL, adjusted pvalue is calculated by taking \code{pvalue} and
dividing it by the maximum number of unique sequences across all samples.}

\item{use_intermediate}{Use saved intermediate files. Used for resuming very
long runs.}

\item{multithread}{TRUE/FALSE or the number of CPU threads to use for multithreading.
Does not work on Windows due to \code{parallel} package implementation.}

\item{verbose}{TRUE/FALSE: display of status messages.}
}
\description{
Uses DADA2 algorithm to denoise trimmed FASTQ files \code{fastq_trimmed}.
DADA2 partitions sequences based on their quality score profiles, using a
\code{pvalue_adjusted} parameter (\code{OMEGA_A} in the paper) as a cutoff
for forming new partitions. (See DADA2 paper
[Callahan et al. Nature Methods 2016] for details.)
If \code{pvalue}
is given (default), then the \code{pvalue_threshold} is calculated by dividing
the \code{pvalue} by the maximum number of unique sequences from all samples.
Otherwise \code{pvalue_threshold} is used directly.
}
\details{
After denoising, the untrimmed parts of sequences are retrieved from
\code{fastq_untrimmed} files and for each partition a consensus sequence
of that part is concatenaded to its end, omitting everything after
(and including) any indel and N.
}
