% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergepairs.R
\name{detect_overlap_length}
\alias{detect_overlap_length}
\title{Auto-detect overlap length in paired-end reads}
\usage{
detect_overlap_length(
  fq_fwd,
  fq_rev,
  min_sim = 0.75,
  match = 5L,
  mismatch = -5L,
  gap_p = -7L,
  rc_reverse = TRUE,
  ncpu = rexmap_option("ncpu"),
  nseqs = 1000,
  nseqs_seed = 42,
  minalnlen_min = 10,
  minalnlen_max = 70,
  minalnlen_step = 5,
  minalnlen_drop_pct = 5,
  force = T,
  verbose = T
)
}
\arguments{
\item{min_sim}{Minimum % similarity in the overlap region to
accept the overlap alignment as valid}

\item{ncpu}{Number of threads}
}
\description{
Tests different minimum overlap lengths for `merge_pairs()`
}
