% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_trimmer.R
\name{detect_pcr_primers}
\alias{detect_pcr_primers}
\title{Auto-detect PCR primers in a merged FASTQ file}
\usage{
detect_pcr_primers(
  fq,
  verbose = T,
  debug = F,
  pr_fwd_maxoff = 35,
  pr_rev_maxoff = 35,
  max_mismatch = 3,
  nseqs = 1000,
  nseqs_seed = 42,
  min_seqs_pct = 5,
  ncpu = 1,
  ncpu_seqs = 4
)
}
\description{
Samples `nseqs` from an input FASTQ file with merged forward and reverse
reads `fq`, then aligns all combinations of primer pairs from RExMapDB.
These can be seen using `rexmap_option('blast_dbs')`.
}
\details{
Returns a most likely hypervariable region as a character string.
}
