% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_blast}
\alias{download_blast}
\title{Download blastn and makeblastdb executables}
\usage{
download_blast(verbose = T, override_os = NULL)
}
\arguments{
\item{verbose}{(default: TRUE) Whether or not to print progress during download.}
}
\description{
Downloads NCBI blastn and makeblastdb executables for the Operating System
running R and copies it to the RExMap package folder. This is executed automatically
if blastn and makeblastdb executables are missing upon attaching RExMap package.
}
