% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Filter and trim reads. Wrapper for dada2::filterAndTrim}
\usage{
filter_and_trim(
  fq_in,
  fq_out,
  maxEE = 2,
  multithread = himap_option("ncpu"),
  verbose = himap_option("verbose"),
  compress = FALSE,
  ...
)
}
\description{
Filter and trim reads. Wrapper for dada2::filterAndTrim
}
