% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{ftquantile}
\alias{ftquantile}
\title{Quantile from frequency table}
\usage{
ftquantile(ft, prob)
}
\description{
Outputs a quantile corresponding to a given probability \code{prob}
from a frequency table \code{ft}. Like the base R code \code{}
}
