% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{himap_option}
\alias{himap_option}
\title{HiMAP options}
\usage{
himap_option(option_names = NULL)
}
\arguments{
\item{option_names}{A string or a vector of strings with available options. If
not given, then the function lists available options.

Options:}
}
\description{
HiMAP options
}
