% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{himap_setoption}
\alias{himap_setoption}
\title{HiMAP set default options}
\usage{
himap_setoption(option_name, value)
}
\arguments{
\item{option_name}{Name of the option to be changed.}

\item{value}{New value.

Options that can be changed:}
}
\description{
HiMAP set default options
}
