% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ie}
\alias{ie}
\title{Shortcut IF/ELSE function}
\usage{
ie(test, yes, no)
}
\description{
if \code{test} evaluates to TRUE, return \code{yes}. Otherwise return \code{no}.
}
\examples{
ie(1==2, '1 equals 2', '1 does not equal 2')
# Output: '1 does not equal 2'
}
