% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{internet_connection}
\alias{internet_connection}
\title{Check if there's an active internet connection}
\usage{
internet_connection(url = "https://cran.r-project.org")
}
\description{
Return TRUE or FALSE
}
