% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{lineagecsv_to_robj}
\alias{lineagecsv_to_robj}
\title{Read lineages*.csv file and convert it to compressed R object}
\usage{
lineagecsv_to_robj(lincsv, out)
}
\description{
This will be used for taxonomy matching.
}
