% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergepairs.R
\name{merge_pairs}
\alias{merge_pairs}
\title{Merge paired-end reads}
\usage{
merge_pairs(
  fq_fwd,
  fq_rev,
  fq_mer,
  min_sim = 0.75,
  min_aln_len = 50,
  match = 5L,
  mismatch = -5L,
  gap_p = -7L,
  rc_reverse = TRUE,
  ncpu = himap_option("ncpu"),
  verbose = FALSE,
  timing = FALSE
)
}
\arguments{
\item{fq_fwd}{A character vector for input forward FASTQ files (can be gzipped).}

\item{fq_rev}{A character vector for input reverse FASTQ files (can be gzipped).}

\item{fq_mer}{A character vector for output merged FASTQ files.}

\item{min_sim}{Minimum similarity to accept an alignment. A floating point number between 0 and 1. Default: 0.75, corresponding to minimum 75\% similarity in the aligned overlap region.}

\item{min_aln_len}{Minimum alignment length. Ignore alignments below this alignment length.}

\item{match}{Score for character match in the alignment (default: 5.}

\item{mismatch}{Score for character mismatch in the alignment (default: -5)}

\item{gap_p}{Score for gap penalty, either gap opening on gap extension (default: -7).}

\item{rc_reverse}{TRUE/FALSE Reverse complement reverse reads? (default: TRUE)}

\item{ncpu}{Number of CPU threads to use for multithreading.}

\item{verbose}{TRUE/FALSE Display of status messages.}

\item{timing}{TRUE/FALSE Time merging.}
}
\description{
Merges reads using overlap version of the Needleman-Wunsch alignment algorithm and calculate
merged quality scores based on the posterior probabilities (cite: Edgar, Fyelberg)
}
