% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergepairs.R
\name{mergeout_to_table}
\alias{mergeout_to_table}
\title{Convert mergestats table to a normal data.table}
\usage{
mergeout_to_table(mergestats)
}
\description{
Convert mergestats table to a normal data.table
}
