% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{osu_dataset_match}
\alias{osu_dataset_match}
\title{Match OSUs across datasets}
\usage{
osu_dataset_match(
  osu_1,
  osu_2,
  dataset_names = c("dataset1", "dataset2"),
  osu_offset = rexmap_option("osu_offset"),
  verbose = T,
  temp_dir = tempdir(),
  blast_verbose = T
)
}
\arguments{
\item{osu_1}{A list of two data tables.

1. OSU abundance data table (output from \code{\link{abundance}}) with
columns:

\code{sample_id | osu_id | osu_count | pctsim | species}\cr

Additionally, this data table can contain an extra \code{dataset} column
which contains a name of that dataset.

2. OSU sequences data table (output from \code{\link{osu_sequences}}) with
columns:

\code{osu_id | species | qseqid | copy_number | pctsim | sequence}\cr}

\item{osu_2}{Same as \code{osu_1} just for a second dataset.}
}
\description{
Match OSUs across datasets
}
