% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{osuab_genuses}
\alias{osuab_genuses}
\title{Parse genus names and strain counts from each OSU species string}
\usage{
osuab_genuses(osuab, ws = "_", split_char = ",")
}
\description{
Parse genus names and strain counts from each OSU species string
}
