% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_trimmer.R
\name{pcr_primer_trimmer}
\alias{pcr_primer_trimmer}
\title{PCR primer trimmer}
\usage{
pcr_primer_trimmer(fq_in, fq_out, region = NULL, pr_fwd = NULL,
  pr_rev = NULL, pr_fwd_maxoff = 10, pr_rev_maxoff = 10,
  return_noprimer = T, ncpu = himap_option("ncpu"), max_mismatch = 2,
  timing = F)
}
\arguments{
\item{fq_in}{A character vector of input file names.}

\item{fq_out}{A character vector of output file names.}

\item{region}{Hypervariable region used. Used to automatically retrieve PCR
primer sequences for trimming. If this parameter is used (not NULL), then
\code{pr_fwd} and \code{pr_rev} are ignored.}

\item{pr_fwd}{Sequence for the \strong{forward} PCR primer (5' -> 3') in the extended
nucleotide code.}

\item{pr_rev}{Sequence for the \strong{reverse} PCR primer (5' -> 3') in the extended
nucleotide code. If you already have 3' -> 5' sequence, use \code{\link{reverse_complement}}
to obtain the 5' -> 3' sequence.}

\item{pr_fwd_maxoff}{Maximum allowed offset from the 5' sequence end for the
forward PCR primer. Primers aligned further into the read ignored.}

\item{pr_rev_maxoff}{Same but for reverse PCR primer, from the 3' end.}

\item{ncpu}{Number of CPU cores (threads) to use for multithreading. By default
all available CPU cores are used (\code{parallel::detectCores()}).}

\item{max_mismatch}{Maximum allowed number of mismatches for each PCR primer,
ignoring any ambiguous nucleotides (not A,C,G or T).}

\item{timing}{Display run-time at the end of trimming (default: FALSE).}
}
\description{
Trims PCR primers from FASTQ sequences.
}
