% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{pool_himap_results}
\alias{pool_himap_results}
\title{Pool multiple HiMAP outputs}
\usage{
pool_himap_results(result_list, dataset_names = NULL, verbose = T)
}
\arguments{
\item{result_list}{A \code{list()} with each element being a list of two
data tables.

1. OSU abundance data table (output from \code{\link{abundance}}) with
columns:

\code{sample_id | osu_id | osu_count | pctsim | species}\cr

Additionally, this data table can contain an extra \code{dataset} column
which contains a name of that dataset.

2. OSU sequences data table (output from \code{\link{osu_sequences}}) with
columns:

\code{osu_id | species | qseqid | copy_number | pctsim | sequence}\cr}

\item{dataset_names}{A character vector containing dataset names for each
element from the \code{result_list} list. If not given (NULL), each abundance
table (first element from each list) should have a 'dataset' column.}

\item{verbose}{TRUE/FALSE: Print progress during dataset pooling.}
}
\description{
Combine HiMAP outputs from multiple datasets into a single table.
}
