% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print_species}
\alias{print_species}
\title{Plot-friendly OSU labels}
\usage{
print_species(x, xsep = ",", xws = "_", ws = " ", sep = ", ",
  include_sp = "single", decreasing = TRUE, show_count = TRUE,
  show_single = FALSE, count_wrap = c("[", "]"), group_genera = TRUE,
  group_genera_sep = "/", trim_genera_len = NULL,
  trim_species_len = NULL, trim_symbol = ".",
  replace_bacterium = TRUE, max_species = NULL,
  multi_species_sub = "spp.", keep_family_strains = FALSE,
  keep_order_strains = FALSE, keep_class_strains = FALSE,
  keep_phylum_strains = FALSE, keep_sp_strains = FALSE,
  keep_sp_strains_symbol = "_", return_single_entry = TRUE,
  debug = FALSE, show_progress = TRUE)
}
\arguments{
\item{x}{(Required) A character vector of strain names.}

\item{xsep}{A character used for separating strain names in `x`.
Default: ','}

\item{xws}{A character used as a whitespace in `x`. Default: '_'}

\item{ws}{A character used as a whitespace in output (return value) of this
function. Default: ' ' (space)}

\item{sep}{A character string used as a species separator in output of this
function. Default: ', ' (comma followed by a space)}

\item{include_sp}{One of {'single', 'always', 'never'}. Controls display of
strains with unassigned species names (having 'sp.' after genus name). Single
keeps these names when this is the single strain in its genus. 'Always' keeps
these names always. 'Never' never keeps them in the output.
For example is the strain assignments are:
   'Blautia_sp._AF19-13LB,Blautia_sp._AF14-40,Fusicatenibacter_saccharivorans,
   Ruminococcus_sp._1,Ruminococcus_gnavus_2'
setting it to `single` (default) will return:
   'Blautia sp., Fusicatenibacter saccharivorans, Ruminococcus gnavus'
since these unnamed Blautia strains are the only strains in Blautia genus and named
strains are prioritized over unnamed strains for genera . If
this is set to never, then the 'sp.' strains are always ommited and the output
will show the best NAMED species only. Setting to always will return:
   'Blautia sp., Fusicatenibacter saccharivorans, Ruminococcus gnavus/sp.'
to indicate that there are more hits in Ruminococcus genus than the only named
one (gnavus).
Default: 'single'}

\item{decreasing}{Sort order of the output species. Default: TRUE sorts the
output starting with the species that had most strain hits.}

\item{show_count}{Display count of strains next to each species name in the
output. Default: TRUE}

\item{show_single}{If show_count == TRUE, this controls whether to display
number 1 for species that match only 1 strain. Default: FALSE (i.e. numbers
are only shown for species mapping to > 1 strain).}

\item{count_wrap}{A character vector of length 2 controling the string
format of the count output, Wraps the count in the two strings, i.e. 3 will
be shown as [3] by default.}

\item{group_genera}{A boolean controling the grouping of species belonging
to the same Genus. Default: TRUE.}

\item{group_genera_sep}{A character symbol used for separating different
species in the same Genus group.}

\item{trim_genera_len}{An integer controling the trimming length for long
Genus names. Any names longer than `trim_genera_len` will be trimmed to
`trim_general_len` followed by a `trim_symbol`.
Default: NULL (do not trim labels)}

\item{trim_species_len}{Same as `trim_genera_len` but for species names
instead of Genus names. Default: NULL (do not trim labels)}

\item{trim_symbol}{A character symbol used after trimmed names. Default: '.'
(A natural choice).}

\item{replace_bacterium}{A boolean controling the handling of strains with
unassigned species names that are not `sp.` but are `bacterium` or `cf.`. If
this is set to TRUE any names ending with `bacterium` and `cf.` are substituted
with `sp.` for consistency. Default: TRUE}

\item{multi_species_sub}{If there are too many species under a single
genus (greater than max_species), then use this string as replacement.
(default: `spp.`)}

\item{keep_family_strains}{Keep strains that do not even have genus assigned, such as
Ruminococcaceae sp. if there are other, strains with named genera.}

\item{keep_sp_strains}{Show full strain names for species names such as
Somegenus_sp._AB01. In this case full strain name is added to the species name}

\item{return_single_entry}{If the input string consists of a single strain name
should we just return that one and ignore other filters? Default: TRUE}
}
\description{
Condense a character vector of strain names to a more readable shorter
string, consisting of species names with optional number of strains matched
for each species.
}
\details{
This function expects input string (or a vector of strings) that consists of strain
names. By default, the strain names are separated by comma, and an underscore is used
as a whitespace symbol (these can be changed: see parameters `xsep`, `xws`).
}
