% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print_strains}
\alias{print_strains}
\title{Strain vector to shorter strain string}
\usage{
print_strains(strains, raw = T)
}
\description{
Condense a character vector of strain names to a more managable
shorter string.
}
