% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{random_sequences}
\alias{random_sequences}
\title{Generate random sequence of given length}
\usage{
random_sequences(len = 50, n = 10, dictionary = c("A", "C", "G",
  "T"))
}
\description{
Generate random sequence of given length
}
