% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_trimmer.R
\name{remove_pcr_primers}
\alias{remove_pcr_primers}
\title{PCR primer trimmer}
\usage{
remove_pcr_primers(
  fq_in,
  fq_out,
  region = NULL,
  pr_fwd = NULL,
  pr_rev = NULL,
  pr_fwd_maxoff = 35,
  pr_rev_maxoff = 35,
  max_mismatch = 2,
  check_rc = FALSE,
  return_noprimer = T,
  ncpu = 1,
  ncpu_sample = rexmap_option("ncpu"),
  verbose = rexmap_option("verbose"),
  overwrite = TRUE
)
}
\arguments{
\item{fq_in}{A character vector of input file names.}

\item{fq_out}{A character vector of output file names.}

\item{region}{Hypervariable region used. Used to automatically retrieve PCR
primer sequences for trimming. If this parameter is used (not NULL), then
\code{pr_fwd} and \code{pr_rev} are ignored.}

\item{pr_fwd}{Sequence for the \strong{forward} PCR primer (5' -> 3') in the extended
nucleotide code.}

\item{pr_rev}{Sequence for the \strong{reverse} PCR primer (5' -> 3') in the extended
nucleotide code. If you already have 3' -> 5' sequence, use \code{\link{reverse_complement}}
to obtain the 5' -> 3' sequence.}

\item{pr_fwd_maxoff}{Maximum allowed offset from the 5' sequence end for the
forward PCR primer. Primers aligned further into the read ignored.}

\item{pr_rev_maxoff}{Same but for reverse PCR primer, from the 3' end.}

\item{max_mismatch}{Maximum allowed number of mismatches for each PCR primer,
ignoring any ambiguous nucleotides (not A,C,G or T).}

\item{check_rc}{Check reverse complements of primer pairs? (default: FALSE)}

\item{ncpu}{Number of CPU cores (threads) to use for multithreading. By default
all available CPU cores are used (\code{parallel::detectCores()}).}

\item{overwrite}{Overwrite target files if they exist (default: TRUE)}
}
\description{
Trims PCR primers from FASTQ sequences.
}
