% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexmap.R
\name{rexmap_db_version}
\alias{rexmap_db_version}
\title{Show HiMAP database version based on the last modification date}
\usage{
rexmap_db_version(region = NULL)
}
\description{
Version == last modified date of the *_unique_variants_R.txt table.
}
