% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexmap.R
\name{rexmap_option}
\alias{rexmap_option}
\title{RExMap options}
\usage{
rexmap_option(option_names = NULL)
}
\arguments{
\item{option_names}{A string or a vector of strings with available options. If
not given, then the function lists available options.}
}
\description{
RExMap options
}
