% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexmap.R
\name{rexmap_setoption}
\alias{rexmap_setoption}
\title{RExMap set default options}
\usage{
rexmap_setoption(option_name, value)
}
\arguments{
\item{option_name}{Name of the option to be changed.}

\item{value}{New value.

Options that can be changed:}
}
\description{
RExMap set default options
}
