% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sampleids_from_filenames}
\alias{sampleids_from_filenames}
\title{Extract sample identifiers from FASTQ filenames}
\usage{
sampleids_from_filenames(filenames, separator = "_")
}
\arguments{
\item{filenames}{A character vector of filenames.}

\item{separator}{A single character used delimiting sample id from the
rest of FASTQ filename.}
}
\description{
Extract sample identifiers from FASTQ filenames
}
