% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexmap.R
\name{sequence_abundance}
\alias{sequence_abundance}
\title{Sequence abundance table}
\usage{
sequence_abundance(
  dada_result,
  remove_bimeras = T,
  collapse_sequences = T,
  verbose = T,
  remove_bimeras_method = "consensus",
  remove_bimeras_oneoff = T,
  fq_prefix_split = ".",
  fq_prefix_split_n = 1,
  ncpu = TRUE,
  debug = F
)
}
\arguments{
\item{dada_result}{dada() result}

\item{remove_bimeras}{Check and remove bimeric reads? (default: TRUE)}

\item{collapse_sequences}{Should we check for sequences that differ up to
shifts and add up the counts? (default: TRUE)}

\item{remove_bimeras_method}{Method to remove bimeras. See ?dada2::removeBimeraDenovo
for more options.}
}
\description{
Return abundances of each sequence from a DADA2 result object.
}
