% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{sequence_length_table}
\alias{sequence_length_table}
\title{Frequency table of sequence lengths}
\usage{
sequence_length_table(fastq_files)
}
\arguments{
\item{fastq_files}{A character vector of FASTQ filenames.}
}
\description{
Generate a frequency table of sequence lengths from FASTQ files.
The resulting table object can be visuaized with base plot function.
}
\examples{


}
