% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summarize_osu_label}
\alias{summarize_osu_label}
\title{Summarize OSU labels}
\usage{
summarize_osu_label(x, show.sp = FALSE, sep = ",")
}
\description{
Shorten sometimes long OSU labels which can list many strains or species.
This removes
}
