% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{untrim}
\alias{untrim}
\title{Un-trim trimmed sequences}
\usage{
untrim(
  dada_res,
  fq_trimmed,
  fq_pretrimmed,
  verbose = himap_option("verbose"),
  ncpu = himap_option("ncpu")
)
}
\description{
Use all pre-trimmed sequences to find the consensus of the trimmed part.
}
\details{
This function modifies dada result \code{dada_res}.
}
