% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{update_database}
\alias{update_database}
\title{Update the RExMap database}
\usage{
update_database(verbose = T, source_repo = "rexmap", update_ref_table = FALSE)
}
\arguments{
\item{verbose}{Verbose output, print status message during execution.}

\item{source_repo}{Download RExMap DB from himap Github repository
(default: 'rexmap') or RExMap db repository ('rexmapdb').}

\item{update_ref_table}{Download and update reference table from the RExMap
repository. Typically this table can be generated automatically on update
when this is set to FALSE (default).}
}
\description{
Updates the database from a RExMap GitHub repository with files pre-generated by
the RExMap authors. For manual database generation see:
https://www.github.com/taolonglab/himapdb .
}
