% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{update_pcr_primers_table}
\alias{update_pcr_primers_table}
\title{Construct the pcr_primers_table after database update
(this is not finished yet)}
\usage{
update_pcr_primers_table(backup_old = T, verbose = T)
}
\arguments{
\item{backup_old}{Whether to backup the current pcr_primers_table before
overwritting. Default: TRUE}

\item{verbose}{Verbose output, print status.}
}
\description{
Automatically generates and verified inst/extdata/pcr_primers_table.txt
file based on the files in database/ folder.
}
