% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_table}
\alias{write_table}
\title{Write data table to file}
\usage{
write_table(table, output, sep = "\\t", verbose = F)
}
\description{
Like base function \code{\link{write.table}} but with normal defaults.
}
