% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{abundance}
\alias{abundance}
\title{Combine BLAST object and a sequence abundance table into OSU table}
\usage{
abundance(abundance_table, blast_object, ncpu = himap_option("ncpu"),
  verbose = T, raw_strains = FALSE)
}
\arguments{
\item{abundance_table}{Sequence abundance table}

\item{blast_object}{BLAST output class, output from \code{\link{blast}} function.}

\item{ncpu}{Integer specifying number of CPU threads to use. This uses R package "parallel"
(TRUE) or to simplify the output when multiple strains of the same species are in the
same OSU (FALSE).}

\item{verbose}{TRUE/FALSE: display status messages}

\item{raw_strains}{TRUE/FALSE: whether to report full strain information for each OSU}
}
\description{
Combine BLAST object and a sequence abundance table into OSU table
}
