% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse sequences that are exact matches, up to shifts and/or length}
\usage{
collapse(ab_in, verbose = himap_option("verbose"))
}
\description{
Collapse sequences that are exact matches, up to shifts and/or length
}
