% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{ftquantile}
\alias{ftquantile}
\title{Quantile from frequency table}
\usage{
ftquantile(ft, prob)
}
\arguments{
\item{ft}{Frequency table of sequence lengths. Output from \code{link{sequence_length_table}}.}

\item{prob}{Probability.}
}
\description{
Outputs a quantile corresponding to a given probability \code{prob}
from a frequency table \code{ft}. We use this to find the minimum sequence length
above which we have certain fraction of reads.
}
\details{
Similar to the base R code \code{\link{quantile}}, but works with
frequency table instead of the raw list.
}
\examples{
# Find minimum sequence length, above which we have 99\% reads
# (this example requires valid FASTQ files in fq_tri variable)
seqlen.ft = sequence_length_table(fq_tri)
ftquantile(seqlen.ft, 0.01)

}
