% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{load_taxonomy}
\alias{load_taxonomy}
\title{Load RExMap taxonomy reference file}
\usage{
load_taxonomy(tax_file = rexmap_option("taxonomy_file"))
}
\arguments{
\item{tax_file}{Full path to the Rdata taxonomy file. Default:
rexmap_option('taxonomy_file').}
}
\description{
This is a NCBI Taxonomy database, exported as an compressed R object.
}
