% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{load_taxonomy}
\alias{load_taxonomy}
\title{Load HiMAP taxonomy reference file}
\usage{
load_taxonomy(tax_file = himap_option("taxonomy_file"))
}
\description{
This is a NCBI Taxonomy database, exported as a tab-delimited file.
}
