% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print_species}
\alias{print_species}
\title{Plot-friendly OSU labels}
\usage{
print_species(x, xsep = ",", xws = "_", ws = " ", sep = ", ",
  include_sp = "single", decreasing = TRUE, show_count = TRUE,
  show_single = FALSE, count_wrap = c("[", "]"), group_genera = TRUE,
  group_genera_sep = "/", trim_genera_len = NULL,
  trim_species_len = NULL, trim_symbol = ".",
  replace_bacterium = TRUE)
}
\arguments{
\item{x}{(Required) A character vector of strain names.}

\item{xsep}{A character used for separating strain names in `x`.
Default: ','}

\item{xws}{A character used as a whitespace in `x`. Default: '_'}

\item{ws}{A character used as a whitespace in output (return value) of this
function. Default: ' ' (space)}

\item{sep}{A character string used as a species separator in output of this
function. Default: ', ' (comma followed by a space)}

\item{include_sp}{One of {'single', 'always', 'never'}. Controls display of
strains with unassigned species names (having 'sp.' after Genus name). Single
keeps these names when this is the single strain in its Genus. 'Always' keeps
these names always. 'Never' never keeps them in the output. Default: 'single'}

\item{decreasing}{Sort order of the output species. Default: TRUE sorts the
output starting with the species that had most strain hits.}

\item{show_count}{Display count of strains next to each species name in the
output. Default: TRUE}

\item{show_single}{If show_count == TRUE, this controls whether to display
number 1 for species that match only 1 strain. Default: FALSE (i.e. numbers
are only shown for species mapping to > 1 strain).}

\item{count_wrap}{A character vector of length 2 controling the string
format of the count output, Wraps the count in the two strings, i.e. 3 will
be shown as [3] by default.}

\item{group_genera}{A boolean controling the grouping of species belonging
to the same Genus. Default: TRUE.}

\item{group_genera_sep}{A character symbol used for separating different
species in the same Genus group.}

\item{trim_genera_len}{An integer controling the trimming length for long
Genus names. Any names longer than `trim_genera_len` will be trimmed to
`trim_general_len` followed by a `trim_symbol`.
Default: NULL (do not trim labels)}

\item{trim_species_len}{Same as `trim_genera_len` but for species names
instead of Genus names. Default: NULL (do not trim labels)}

\item{trim_symbol}{A character symbol used after trimmed names. Default: '.'
(A natural choice).}

\item{replace_bacterium}{A boolean controling the handling of strains with
unassigned species names that are not `sp.` but are `bacterium` or `cf.`. If
this is set to TRUE any names with `bacterium` and `cf.` are substituted
with `sp.` for ease of display. Default: TRUE}
}
\description{
Condense a character vector of strain names to a more readable shorter
string, consisting of species names with optional number of strains matched
for each species.
}
