% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himap.R
\name{sequence_abundance}
\alias{sequence_abundance}
\title{Return abundances of each sequence from DADA2 result object}
\usage{
sequence_abundance(dada_result, remove_bimeras = T,
  collapse_sequences = T, verbose = T,
  remove_bimeras_method = "consensus", remove_bimeras_oneoff = T,
  fq_prefix_split = ".")
}
\arguments{
\item{dada_result}{dada() result}

\item{remove_bimeras}{Check and remove bimeric reads? (default: TRUE)}

\item{collapse_sequences}{Should we check for sequences that differ up to
shifts and add up the counts? (default: TRUE)}

\item{remove_bimeras_method}{Method to remove bimeras. See ?dada2::removeBimeraDenovo
for more options.}
}
\description{
Return abundances of each sequence from DADA2 result object
}
