% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Generate taxonomy for each OSU}
\usage{
taxonomy(osu_abundance_table, verbose = FALSE)
}
\arguments{
\item{osu_abundance_table}{Data table with OSU abundances. Output from abundance().sssss}
}
\description{
Generate taxonomy for each OSU
}
