% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{update_database}
\alias{update_database}
\title{Update the HiMAP database}
\usage{
update_database(verbose = T, source_repo = "himap",
  update_ref_table = FALSE)
}
\arguments{
\item{verbose}{Verbose output, print status message during execution.}

\item{source_repo}{Download HiMAP DB from himap Github repository
(default: 'himap') or HiMAP db repository ('himapdb').}

\item{update_ref_table}{Download and update reference table from the HiMAP
repository. Typically this table can be generated automatically on update
when this is set to FALSE (default).}
}
\description{
Updates the database from a HiMAP GitHub repository with files pre-generated by
the HiMAP authors. For manual database generation see:
https://www.github.com/taolonglab/himapdb .
}
